restructure AngularKet
