+++ src/rydstate/angular/__init__.py
@@ -1,14 +1,10 @@
 from rydstate.angular import utils
 from rydstate.angular.angular_ket import AngularKetFJ, AngularKetJJ, AngularKetLS
-from rydstate.angular.angular_ket_base import AngularKetBaseFJ, AngularKetBaseJJ, AngularKetBaseLS
 from rydstate.angular.angular_state import AngularState
 from rydstate.angular.core_ket import CoreKet
 from rydstate.angular.utils import NotSet
 
 __all__ = [
-    "AngularKetBaseFJ",
-    "AngularKetBaseJJ",
-    "AngularKetBaseLS",
     "AngularKetFJ",
     "AngularKetJJ",
     "AngularKetLS",
