+++ src/rydstate/species/mqdt/yb171.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import numpy as np
 
 from rydstate.angular.angular_ket import AngularKetFJ, AngularKetJJ, AngularKetLS
-from rydstate.angular.angular_ket_dummy import AngularKetDummyFJ
 from rydstate.angular.core_ket import CoreKet
 from rydstate.angular.utils import Unknown
 from rydstate.species.mqdt.fmodel import FModel
@@ -56,22 +55,22 @@ class Yb171_S05_HighN(FModel):
 
     inner_channels = [
         AngularKetLS(l_c=0, l_r=0, l_tot=0, s_tot=0, j_tot=0, f_tot=0.5, species="Yb171"),  # "6sns 1S0"
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
         AngularKetLS(l_c=1, l_r=1, l_tot=0, s_tot=0, j_tot=0, f_tot=0.5, species="Yb171"),  # "6pnp 1S0"
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
         AngularKetLS(l_c=1, l_r=1, l_tot=1, s_tot=1, j_tot=0, f_tot=0.5, species="Yb171"),  # "6pnp 3P0"
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
         AngularKetLS(l_c=0, l_r=0, l_tot=0, s_tot=1, j_tot=1, f_tot=0.5, species="Yb171"),  # "6sns 3S1"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=0, j_c=0.5, f_c=0, j_r=0.5, f_tot=0.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
         AngularKetFJ(l_c=1, l_r=1, j_c=1.5, f_c=1, j_r=1.5, f_tot=0.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(
-            l_c=1, l_r=1, j_c=0.5, j_r=0.5, f_tot=0.5, name="f_c unknown", species="Yb171"
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(
+            l_c=1, l_r=1, j_c=0.5, j_r=0.5, f_tot=0.5, allow_unknown=True, name="f_c unknown", species="Yb171"
         ),  # just add two states f_c = 0 and f_c = 1?
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=0, j_c=0.5, f_c=1, j_r=0.5, f_tot=0.5, species="Yb171"),
     ]
 
@@ -135,22 +134,22 @@ class Yb171_P05_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=0, j_tot=1, f_tot=0.5, species="Yb171"),  # "6snp 1P1"
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=1, f_tot=0.5, species="Yb171"),  # "6snp 3P1"
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl c", species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl d", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb171"),
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=0, f_tot=0.5, species="Yb171"),  # "6snp 3P0"
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl e", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb171"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=1, j_r=1.5, f_tot=0.5, species="Yb171"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=1, j_r=0.5, f_tot=0.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl c", species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl d", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=0, j_r=0.5, f_tot=0.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl e", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb171"),
     ]
 
     eigen_quantum_defects = [
@@ -188,27 +187,27 @@ class Yb171_P15_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=0, j_tot=1, f_tot=1.5, species="Yb171"),  # "6snp 1P1"
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=1, f_tot=1.5, species="Yb171"),  # "6snp 3P1"
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl c", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl d", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb171"),
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=2, f_tot=1.5, species="Yb171"),  # "6snp 3P2"
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl e", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl f", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl g", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb171"),
         AngularKetLS(l_c=0, l_r=3, l_tot=3, s_tot=1, j_tot=2, f_tot=1.5, species="Yb171"),  # "6snf 3F2"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=1, j_r=1.5, f_tot=1.5, species="Yb171"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=1, j_r=0.5, f_tot=1.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl c", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl d", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=0, j_r=1.5, f_tot=1.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl e", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl f", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl g", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=3, j_c=0.5, f_c=1, j_r=2.5, f_tot=1.5, species="Yb171"),
     ]
 
@@ -272,17 +271,17 @@ class Yb171_D15_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=0, j_tot=2, f_tot=1.5, species="Yb171"),  # "6snd 1D2"
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=1, j_tot=2, f_tot=1.5, species="Yb171"),  # "6snd 3D2"
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
         AngularKetLS(l_c=1, l_r=1, l_tot=2, s_tot=0, j_tot=2, f_tot=1.5, species="Yb171"),  # "6pnp 1D2"
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=1, j_tot=1, f_tot=1.5, species="Yb171"),  # "6snd 3D1"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=1, j_r=2.5, f_tot=1.5, species="Yb171"),
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=1, j_r=1.5, f_tot=1.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(l_c=1, l_r=1, f_tot=1.5, name="j_c and j_r unknown", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(l_c=1, l_r=1, f_tot=1.5, allow_unknown=True, name="j_c and j_r unknown", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=0, j_r=1.5, f_tot=1.5, species="Yb171"),
     ]
     manual_frame_transformation_outer_inner = np.array(
@@ -324,17 +323,17 @@ class Yb171_D25_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=0, j_tot=2, f_tot=2.5, species="Yb171"),  # "6snd 1D2"
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=1, j_tot=2, f_tot=2.5, species="Yb171"),  # "6snd 3D2"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
         AngularKetLS(l_c=1, l_r=1, l_tot=2, s_tot=0, j_tot=2, f_tot=2.5, species="Yb171"),  # "6pnp 1D2"
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=1, j_tot=3, f_tot=2.5, species="Yb171"),  # "6snd 3D3"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=1, j_r=2.5, f_tot=2.5, species="Yb171"),
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=1, j_r=1.5, f_tot=2.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(l_c=1, l_r=1, f_tot=2.5, name="j_c and j_r unknown", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(l_c=1, l_r=1, f_tot=2.5, allow_unknown=True, name="j_c and j_r unknown", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=0, j_r=2.5, f_tot=2.5, species="Yb171"),
     ]
     manual_frame_transformation_outer_inner = np.array(
@@ -396,29 +395,29 @@ class Yb171_F25_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=3, l_tot=3, s_tot=0, j_tot=3, f_tot=2.5, species="Yb171"),  # "6snf 1F3"
         AngularKetLS(l_c=0, l_r=3, l_tot=3, s_tot=1, j_tot=3, f_tot=2.5, species="Yb171"),  # "6snf 3F3"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl c", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl d", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl e", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb171"),
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=2, f_tot=2.5, species="Yb171"),  # "6snf 3P2"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl f", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl g", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl h", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl h", species="Yb171"),
         AngularKetLS(l_c=0, l_r=3, l_tot=3, s_tot=1, j_tot=2, f_tot=2.5, species="Yb171"),  # "6snf 3F2"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=3, j_c=0.5, f_c=1, j_r=3.5, f_tot=2.5, species="Yb171"),
         AngularKetFJ(l_c=0, l_r=3, j_c=0.5, f_c=1, j_r=2.5, f_tot=2.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl c", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl d", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl e", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=1, j_r=1.5, f_tot=2.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl f", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl g", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl h", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl h", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=3, j_c=0.5, f_c=0, j_r=2.5, f_tot=2.5, species="Yb171"),
     ]
 
@@ -465,21 +464,21 @@ class Yb171_F35_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=3, l_tot=3, s_tot=0, j_tot=3, f_tot=3.5, species="Yb171"),  # "6snf 1F3"
         AngularKetLS(l_c=0, l_r=3, l_tot=3, s_tot=1, j_tot=3, f_tot=3.5, species="Yb171"),  # "6snf 3F3"
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl c", species="Yb171"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl d", species="Yb171"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl e", species="Yb171"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb171"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb171"),
         AngularKetLS(l_c=0, l_r=3, l_tot=3, s_tot=1, j_tot=4, f_tot=3.5, species="Yb171"),  # "6snf 3F4"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=3, j_c=0.5, f_c=1, j_r=3.5, f_tot=3.5, species="Yb171"),
         AngularKetFJ(l_c=0, l_r=3, j_c=0.5, f_c=1, j_r=2.5, f_tot=3.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl c", species="Yb171"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl d", species="Yb171"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl e", species="Yb171"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb171"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=3, j_c=0.5, f_c=0, j_r=3.5, f_tot=3.5, species="Yb171"),
     ]
 
@@ -638,22 +637,22 @@ class Yb171_S05_LowN(FModel):
 
     inner_channels = [
         AngularKetLS(l_c=0, l_r=0, l_tot=0, s_tot=0, j_tot=0, f_tot=0.5, species="Yb171"),  # "6sns 1S0"
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
         AngularKetLS(l_c=1, l_r=1, l_tot=0, s_tot=0, j_tot=0, f_tot=0.5, species="Yb171"),  # "6pnp 1S0"
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
         AngularKetLS(l_c=1, l_r=1, l_tot=1, s_tot=1, j_tot=0, f_tot=0.5, species="Yb171"),  # "6pnp 3P0"
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
         AngularKetLS(l_c=0, l_r=0, l_tot=0, s_tot=1, j_tot=1, f_tot=0.5, species="Yb171"),  # "6sns 3S1"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=0, j_c=0.5, f_c=0, j_r=0.5, f_tot=0.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
         AngularKetFJ(l_c=1, l_r=1, j_c=1.5, f_c=1, j_r=1.5, f_tot=0.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(
-            l_c=1, l_r=1, j_c=0.5, j_r=0.5, f_tot=0.5, name="f_c unknown", species="Yb171"
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(
+            l_c=1, l_r=1, j_c=0.5, j_r=0.5, f_tot=0.5, allow_unknown=True, name="f_c unknown", species="Yb171"
         ),  # just add two states f_c = 0 and f_c = 1?
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=0, j_c=0.5, f_c=1, j_r=0.5, f_tot=0.5, species="Yb171"),
     ]
 
@@ -800,26 +799,26 @@ class Yb171_P15_LowN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=0, j_tot=1, f_tot=1.5, species="Yb171"),  # "6snp 1P1"
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=1, f_tot=1.5, species="Yb171"),  # "6snp 3P1"
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl c", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl d", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb171"),
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=2, f_tot=1.5, species="Yb171"),  # "6snp 3P2"
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl e", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl f", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl g", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb171"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=1, j_r=1.5, f_tot=1.5, species="Yb171"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=1, j_r=0.5, f_tot=1.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl c", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl d", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=0, j_r=1.5, f_tot=1.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl e", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl f", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl g", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb171"),
     ]
 
     eigen_quantum_defects = [
@@ -879,15 +878,15 @@ class Yb171_P25_LowN(FModel):
 
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=2, f_tot=2.5, species="Yb171"),  # "6snp 3P2"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=1, j_r=1.5, f_tot=2.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl c", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb171"),
     ]
 
     eigen_quantum_defects = [
@@ -933,17 +932,17 @@ class Yb171_D15_LowN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=0, j_tot=2, f_tot=1.5, species="Yb171"),  # "6snd 1D2"
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=1, j_tot=2, f_tot=1.5, species="Yb171"),  # "6snd 3D2"
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
         AngularKetLS(l_c=1, l_r=1, l_tot=2, s_tot=0, j_tot=2, f_tot=1.5, species="Yb171"),  # "6pnp 1D2"
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=1, j_tot=1, f_tot=1.5, species="Yb171"),  # "6snd 3D1"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=1, j_r=2.5, f_tot=1.5, species="Yb171"),
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=1, j_r=1.5, f_tot=1.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(l_c=1, l_r=1, f_tot=1.5, name="j_c and j_r unknown", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(l_c=1, l_r=1, f_tot=1.5, allow_unknown=True, name="j_c and j_r unknown", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=0, j_r=1.5, f_tot=1.5, species="Yb171"),
     ]
     manual_frame_transformation_outer_inner = np.array(
@@ -985,17 +984,17 @@ class Yb171_D25_LowN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=0, j_tot=2, f_tot=2.5, species="Yb171"),  # "6snd 1D2"
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=1, j_tot=2, f_tot=2.5, species="Yb171"),  # "6snd 3D2"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
         AngularKetLS(l_c=1, l_r=1, l_tot=2, s_tot=0, j_tot=2, f_tot=2.5, species="Yb171"),  # "6pnp 1D2"
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=1, j_tot=3, f_tot=2.5, species="Yb171"),  # "6snd 3D3"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=1, j_r=2.5, f_tot=2.5, species="Yb171"),
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=1, j_r=1.5, f_tot=2.5, species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb171"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb171"),
-        AngularKetDummyFJ(l_c=1, l_r=1, f_tot=2.5, name="j_c and j_r unknown", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb171"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb171"),
+        AngularKetFJ(l_c=1, l_r=1, f_tot=2.5, allow_unknown=True, name="j_c and j_r unknown", species="Yb171"),
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, f_c=0, j_r=2.5, f_tot=2.5, species="Yb171"),
     ]
     manual_frame_transformation_outer_inner = np.array(
