+++ src/rydstate/species/mqdt/yb173.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import numpy as np
 
 from rydstate.angular.angular_ket import AngularKetFJ, AngularKetLS
-from rydstate.angular.angular_ket_dummy import AngularKetDummyFJ
 from rydstate.angular.core_ket import CoreKet
 from rydstate.angular.utils import Unknown
 from rydstate.species.mqdt.fmodel import FModel
@@ -74,22 +73,22 @@ class Yb173_S25_HighN(FModel):
 
     inner_channels = [
         AngularKetLS(l_c=0, l_r=0, l_tot=0, s_tot=0, j_tot=0, f_tot=2.5, species="Yb173"),  # "6sns 1S0"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
         AngularKetLS(l_c=1, l_r=1, l_tot=0, s_tot=0, j_tot=0, f_tot=2.5, species="Yb173"),  # "6pnp 1S0"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
         AngularKetLS(l_c=1, l_r=1, l_tot=1, s_tot=1, j_tot=0, f_tot=2.5, species="Yb173"),  # "6pnp 3P0"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
         AngularKetLS(l_c=0, l_r=0, l_tot=0, s_tot=1, j_tot=1, f_tot=2.5, species="Yb173"),  # "6sns 3S1"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=0, j_c=0.5, f_c=2, j_r=0.5, f_tot=2.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
         AngularKetFJ(l_c=1, l_r=1, j_c=1.5, f_c=3, j_r=1.5, f_tot=2.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(
-            l_c=1, l_r=1, j_c=0.5, j_r=0.5, f_tot=2.5, name="f_c unknown", species="Yb173"
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(
+            l_c=1, l_r=1, j_c=0.5, j_r=0.5, f_tot=2.5, allow_unknown=True, name="f_c unknown", species="Yb173"
         ),  # just add two states f_c = 2 and f_c = 3?
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
         AngularKetFJ(l_c=0, l_r=0, j_c=0.5, f_c=3, j_r=0.5, f_tot=2.5, species="Yb173"),
     ]
     manual_frame_transformation_outer_inner = np.array(
@@ -152,15 +151,15 @@ class Yb173_P05_HighN(FModel):
 
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=2, f_tot=0.5, species="Yb173"),  # "6snp 3P2"
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl a", species="Yb173"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=2, j_r=1.5, f_tot=0.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl a", species="Yb173"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(f_tot=0.5, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=0.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
     ]
 
     eigen_quantum_defects = [
@@ -186,26 +185,26 @@ class Yb173_P15_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=0, j_tot=1, f_tot=1.5, species="Yb173"),  # "6snp 1P1"
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=1, f_tot=1.5, species="Yb173"),  # "6snp 3P1"
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl a", species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl c", species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl d", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb173"),
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=2, f_tot=1.5, species="Yb173"),  # "6snp 3P2"
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl e", species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl f", species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl g", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb173"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=2, j_r=0.5, f_tot=1.5, species="Yb173"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=2, j_r=1.5, f_tot=1.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl a", species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl c", species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl d", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb173"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=3, j_r=1.5, f_tot=1.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl e", species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl f", species="Yb173"),
-        AngularKetDummyFJ(f_tot=1.5, name="4f13 5d 6snl g", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb173"),
+        AngularKetFJ(f_tot=1.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb173"),
     ]
 
     eigen_quantum_defects = [
@@ -246,30 +245,30 @@ class Yb173_P25_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=0, j_tot=1, f_tot=2.5, species="Yb173"),  # "6snp 1P1"
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=1, f_tot=2.5, species="Yb173"),  # "6snp 3P1"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl c", species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl d", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb173"),
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=0, f_tot=2.5, species="Yb173"),  # "6snp 3P0"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl e", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb173"),
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=2, f_tot=2.5, species="Yb173"),  # "6snp 3P2"
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl f", species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl g", species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl h", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl h", species="Yb173"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=2, j_r=0.5, f_tot=2.5, species="Yb173"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=2, j_r=1.5, f_tot=2.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl a", species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl c", species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl d", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb173"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=3, j_r=0.5, f_tot=2.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl e", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb173"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=3, j_r=1.5, f_tot=2.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl f", species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl g", species="Yb173"),
-        AngularKetDummyFJ(f_tot=2.5, name="4f13 5d 6snl h", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb173"),
+        AngularKetFJ(f_tot=2.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl h", species="Yb173"),
     ]
 
     eigen_quantum_defects = [
@@ -313,26 +312,26 @@ class Yb173_P35_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=0, j_tot=1, f_tot=3.5, species="Yb173"),  # "6snp 1P1"
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=1, f_tot=3.5, species="Yb173"),  # "6snp 3P1"
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl a", species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl c", species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl d", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb173"),
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=2, f_tot=3.5, species="Yb173"),  # "6snp 3P2"
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl e", species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl f", species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl g", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb173"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=2, j_r=1.5, f_tot=3.5, species="Yb173"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=3, j_r=0.5, f_tot=3.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl a", species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl c", species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl d", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb173"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=3, j_r=1.5, f_tot=3.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl e", species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl f", species="Yb173"),
-        AngularKetDummyFJ(f_tot=3.5, name="4f13 5d 6snl g", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl f", species="Yb173"),
+        AngularKetFJ(f_tot=3.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl g", species="Yb173"),
     ]
 
     eigen_quantum_defects = [
@@ -372,15 +371,15 @@ class Yb173_P45_HighN(FModel):
 
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=2, f_tot=4.5, species="Yb173"),  # "6snp 3P2"
-        AngularKetDummyFJ(f_tot=4.5, name="4f13 5d 6snl a", species="Yb173"),
-        AngularKetDummyFJ(f_tot=4.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(f_tot=4.5, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=4.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=4.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=4.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, f_c=3, j_r=1.5, f_tot=4.5, species="Yb173"),
-        AngularKetDummyFJ(f_tot=4.5, name="4f13 5d 6snl a", species="Yb173"),
-        AngularKetDummyFJ(f_tot=4.5, name="4f13 5d 6snl b", species="Yb173"),
-        AngularKetDummyFJ(f_tot=4.5, name="4f13 5d 6snl c", species="Yb173"),
+        AngularKetFJ(f_tot=4.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb173"),
+        AngularKetFJ(f_tot=4.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb173"),
+        AngularKetFJ(f_tot=4.5, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb173"),
     ]
 
     eigen_quantum_defects = [
