+++ src/rydstate/species/mqdt/yb174.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import numpy as np
 
 from rydstate.angular.angular_ket import AngularKetFJ, AngularKetJJ, AngularKetLS
-from rydstate.angular.angular_ket_dummy import AngularKetDummyFJ
 from rydstate.angular.core_ket import CoreKet
 from rydstate.angular.utils import Unknown
 from rydstate.species.mqdt.fmodel import FModel
@@ -55,19 +54,19 @@ class Yb174_S0_HighN(FModel):
 
     inner_channels = [
         AngularKetLS(l_c=0, l_r=0, l_tot=0, s_tot=0, j_tot=0, species="Yb174"),  # "6sns 1S0"
-        AngularKetDummyFJ(f_tot=0, name="4f13 5d 6snl a", species="Yb174"),
+        AngularKetFJ(f_tot=0, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb174"),
         AngularKetLS(l_c=1, l_r=1, l_tot=0, s_tot=0, j_tot=0, species="Yb174"),  # "6pnp 1S0"
-        AngularKetDummyFJ(f_tot=0, name="4f13 5d 6snl b", species="Yb174"),
+        AngularKetFJ(f_tot=0, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb174"),
         AngularKetLS(l_c=1, l_r=1, l_tot=1, s_tot=1, j_tot=0, species="Yb174"),  # "6pnp 3P0"
-        AngularKetDummyFJ(f_tot=0, name="4f13 5d 6snl c", species="Yb174"),
+        AngularKetFJ(f_tot=0, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb174"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=0, j_c=0.5, j_r=0.5, f_tot=0, species="Yb174"),
-        AngularKetDummyFJ(f_tot=0, name="4f13 5d 6snl a", species="Yb174"),
+        AngularKetFJ(f_tot=0, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb174"),
         AngularKetFJ(l_c=1, l_r=1, j_c=1.5, j_r=1.5, f_tot=0, species="Yb174"),
-        AngularKetDummyFJ(f_tot=0, name="4f13 5d 6snl b", species="Yb174"),
+        AngularKetFJ(f_tot=0, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb174"),
         AngularKetFJ(l_c=1, l_r=1, j_c=0.5, j_r=0.5, f_tot=0, species="Yb174"),
-        AngularKetDummyFJ(f_tot=0, name="4f13 5d 6snl c", species="Yb174"),
+        AngularKetFJ(f_tot=0, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb174"),
     ]
 
     eigen_quantum_defects = [
@@ -117,11 +116,11 @@ class Yb174_P0_HighN(FModel):
 
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=0, species="Yb174"),  # "6snp 3P0"
-        AngularKetDummyFJ(f_tot=0, name="4f13 5d 6snd", species="Yb174"),
+        AngularKetFJ(f_tot=0, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snd", species="Yb174"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, j_r=0.5, f_tot=0, species="Yb174"),
-        AngularKetDummyFJ(f_tot=0, name="4f13 5d 6snd", species="Yb174"),
+        AngularKetFJ(f_tot=0, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snd", species="Yb174"),
     ]
 
     eigen_quantum_defects = [
@@ -143,18 +142,18 @@ class Yb174_P1_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=0, j_tot=1, species="Yb174"),  # "6snp 1P1"
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=1, species="Yb174"),  # "6snp 3P1"
-        AngularKetDummyFJ(f_tot=1, name="4f13 5d 6snl a", species="Yb174"),
-        AngularKetDummyFJ(f_tot=1, name="4f13 5d 6snl b", species="Yb174"),
-        AngularKetDummyFJ(f_tot=1, name="4f13 5d 6snl c", species="Yb174"),
-        AngularKetDummyFJ(f_tot=1, name="4f13 5d 6snl d", species="Yb174"),
+        AngularKetFJ(f_tot=1, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb174"),
+        AngularKetFJ(f_tot=1, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb174"),
+        AngularKetFJ(f_tot=1, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb174"),
+        AngularKetFJ(f_tot=1, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb174"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, j_r=1.5, f_tot=1, species="Yb174"),
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, j_r=0.5, f_tot=1, species="Yb174"),
-        AngularKetDummyFJ(f_tot=1, name="4f13 5d 6snl a", species="Yb174"),
-        AngularKetDummyFJ(f_tot=1, name="4f13 5d 6snl b", species="Yb174"),
-        AngularKetDummyFJ(f_tot=1, name="4f13 5d 6snl c", species="Yb174"),
-        AngularKetDummyFJ(f_tot=1, name="4f13 5d 6snl d", species="Yb174"),
+        AngularKetFJ(f_tot=1, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb174"),
+        AngularKetFJ(f_tot=1, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb174"),
+        AngularKetFJ(f_tot=1, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb174"),
+        AngularKetFJ(f_tot=1, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb174"),
     ]
 
     eigen_quantum_defects = [
@@ -187,15 +186,15 @@ class Yb174_P2_HighN(FModel):
 
     inner_channels = [
         AngularKetLS(l_c=0, l_r=1, l_tot=1, s_tot=1, j_tot=2, species="Yb174"),  # "6snp 3P2"
-        AngularKetDummyFJ(f_tot=2, name="4f13 5d 6snl a", species="Yb174"),
-        AngularKetDummyFJ(f_tot=2, name="4f13 5d 6snl b", species="Yb174"),
-        AngularKetDummyFJ(f_tot=2, name="4f13 5d 6snl c", species="Yb174"),
+        AngularKetFJ(f_tot=2, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb174"),
+        AngularKetFJ(f_tot=2, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb174"),
+        AngularKetFJ(f_tot=2, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb174"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=1, j_c=0.5, j_r=1.5, f_tot=2, species="Yb174"),
-        AngularKetDummyFJ(f_tot=2, name="4f13 5d 6snl a", species="Yb174"),
-        AngularKetDummyFJ(f_tot=2, name="4f13 5d 6snl b", species="Yb174"),
-        AngularKetDummyFJ(f_tot=2, name="4f13 5d 6snl c", species="Yb174"),
+        AngularKetFJ(f_tot=2, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb174"),
+        AngularKetFJ(f_tot=2, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb174"),
+        AngularKetFJ(f_tot=2, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb174"),
     ]
 
     eigen_quantum_defects = [
@@ -241,17 +240,17 @@ class Yb174_D2_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=0, j_tot=2, species="Yb174"),  # "6snd 1D2"
         AngularKetLS(l_c=0, l_r=2, l_tot=2, s_tot=1, j_tot=2, species="Yb174"),  # "6snd 3D2"
-        AngularKetDummyFJ(f_tot=2, name="4f13 5d 6snl a", species="Yb174"),
-        AngularKetDummyFJ(f_tot=2, name="4f13 5d 6snl b", species="Yb174"),
+        AngularKetFJ(f_tot=2, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb174"),
+        AngularKetFJ(f_tot=2, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb174"),
         AngularKetLS(l_c=1, l_r=1, l_tot=2, s_tot=0, j_tot=2, species="Yb174"),  # "6pnp 1D2"
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, j_r=2.5, f_tot=2, species="Yb174"),
         AngularKetFJ(l_c=0, l_r=2, j_c=0.5, j_r=1.5, f_tot=2, species="Yb174"),
-        AngularKetDummyFJ(f_tot=2, name="4f13 5d 6snl a", species="Yb174"),
-        AngularKetDummyFJ(f_tot=2, name="4f13 5d 6snl b", species="Yb174"),
-        AngularKetDummyFJ(
-            l_c=1, l_r=1, f_tot=2, name="j_c and j_r unknown", species="Yb174"
+        AngularKetFJ(f_tot=2, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb174"),
+        AngularKetFJ(f_tot=2, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb174"),
+        AngularKetFJ(
+            l_c=1, l_r=1, f_tot=2, allow_unknown=True, name="j_c and j_r unknown", species="Yb174"
         ),  # Jc could be either 1/2 or 3/2
     ]
     manual_frame_transformation_outer_inner = np.array(
@@ -331,20 +330,20 @@ class Yb174_F3_HighN(FModel):
     inner_channels = [
         AngularKetLS(l_c=0, l_r=3, l_tot=3, s_tot=0, j_tot=3, species="Yb174"),  # "6snf 1F3"
         AngularKetLS(l_c=0, l_r=3, l_tot=3, s_tot=1, j_tot=3, species="Yb174"),  # "6snf 3F3"
-        AngularKetDummyFJ(f_tot=3, name="4f13 5d 6snl a", species="Yb174"),
-        AngularKetDummyFJ(f_tot=3, name="4f13 5d 6snl b", species="Yb174"),
-        AngularKetDummyFJ(f_tot=3, name="4f13 5d 6snl c", species="Yb174"),
-        AngularKetDummyFJ(f_tot=3, name="4f13 5d 6snl d", species="Yb174"),
-        AngularKetDummyFJ(f_tot=3, name="4f13 5d 6snl e", species="Yb174"),
+        AngularKetFJ(f_tot=3, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb174"),
+        AngularKetFJ(f_tot=3, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb174"),
+        AngularKetFJ(f_tot=3, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb174"),
+        AngularKetFJ(f_tot=3, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb174"),
+        AngularKetFJ(f_tot=3, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb174"),
     ]
     outer_channels = [
         AngularKetFJ(l_c=0, l_r=3, j_c=0.5, j_r=3.5, f_tot=3, species="Yb174"),
         AngularKetFJ(l_c=0, l_r=3, j_c=0.5, j_r=2.5, f_tot=3, species="Yb174"),
-        AngularKetDummyFJ(f_tot=3, name="4f13 5d 6snl a", species="Yb174"),
-        AngularKetDummyFJ(f_tot=3, name="4f13 5d 6snl b", species="Yb174"),
-        AngularKetDummyFJ(f_tot=3, name="4f13 5d 6snl c", species="Yb174"),
-        AngularKetDummyFJ(f_tot=3, name="4f13 5d 6snl d", species="Yb174"),
-        AngularKetDummyFJ(f_tot=3, name="4f13 5d 6snl e", species="Yb174"),
+        AngularKetFJ(f_tot=3, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl a", species="Yb174"),
+        AngularKetFJ(f_tot=3, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl b", species="Yb174"),
+        AngularKetFJ(f_tot=3, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl c", species="Yb174"),
+        AngularKetFJ(f_tot=3, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl d", species="Yb174"),
+        AngularKetFJ(f_tot=3, l_c=Unknown, allow_unknown=True, name="4f13 5d 6snl e", species="Yb174"),
     ]
 
     eigen_quantum_defects = [
